pyproject.toml
[project]
name = "cli-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.13.0",
    "gradio>=6.8.0",
    "python-dotenv>=1.2.1",
]
